@@ -16,12 +16,12 @@
     <packaging>jar</packaging>
     <name>LangChain4j integration with Neo4j Vector Index</name>
 
-    <description>Requires Java 8</description>
+    <description>Requires Java 17</description>
 
     <properties>
         <testcontainers.version>1.19.0</testcontainers.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
     
@@ -36,7 +36,7 @@
         <dependency>
             <groupId>org.neo4j.driver</groupId>
             <artifactId>neo4j-java-driver</artifactId>
-            <version>4.4.12</version>
+            <version>5.15.0</version>
         </dependency>
 
         <dependency>
